# swagger.yaml
openapi: 3.0.0
paths:
...
  /permissions:
    get:
      operationId: getPermissions
      parameters:
        - name: pageSize
          required: false
          in: query
          description: Number of items per page
          schema:
            example: 20
            type: number
        - name: pageIndex
          required: false
          in: query
          description: Page index (starts from 0)
          schema:
            example: 0
            type: number
        - name: order
          required: false
          in: query
          description: Sort order
          schema:
            example: asc
            type: string
            enum:
              - asc
              - desc
        - name: orderBy
          required: false
          in: query
          description: Field to order by
          schema:
            example: createdAt
            type: string
            enum:
              - createdAt
              - description
              - name
              - updatedAt
        - name: keyword
          required: false
          in: query
          description: Search keyword
          schema:
            example: admin
            type: string
        - name: Authorization
          in: header
          description: Bearer auth token
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Retrieve a list of permissions with pagination.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/PageDto"
                  - properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/PermissionWithRolesResponseDto"
                    required:
                      - data
      summary: Get a list of permissions
      tags: &a3
        - Permissions
    post:
      operationId: createPermission
      parameters:
        - name: Authorization
          in: header
          description: Bearer auth token
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePermissionRequestDto"
      responses:
        "200":
          description: Create a new permission with associated roles.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PermissionWithRolesResponseDto"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                example:
                  statusCode: 400
                  message:
                    - field: field
                      message: message
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                example:
                  statusCode: 401
                  message: Unauthorized.
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                example:
                  statusCode: 403
                  message: Forbidden.
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                example:
                  statusCode: 404
                  message: Not Found.
        "422":
          description: Unprocessable Entity
          content:
            application/json:
              schema:
                example:
                  statusCode: 422
                  message: Unprocessable content.
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                example:
                  statusCode: 500
                  message: Internal server error.
      summary: Create a new permission
      tags: *a3
  /permissions/{id}:
    get:
      operationId: getPermissionById
      parameters:
        - name: id
          required: true
          in: path
          description: Permission ID (UUID)
          schema:
            example: 123e4567-e89b-12d3-a456-426614174000
            type: string
        - name: Authorization
          in: header
          description: Bearer auth token
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Retrieve a specific permission by its ID.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PermissionWithRolesResponseDto"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                example:
                  statusCode: 400
                  message:
                    - field: field
                      message: message
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                example:
                  statusCode: 401
                  message: Unauthorized.
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                example:
                  statusCode: 403
                  message: Forbidden.
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                example:
                  statusCode: 404
                  message: Not Found.
        "422":
          description: Unprocessable Entity
          content:
            application/json:
              schema:
                example:
                  statusCode: 422
                  message: Unprocessable content.
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                example:
                  statusCode: 500
                  message: Internal server error.
      summary: Get a permission by ID
      tags: *a3
    put:
      operationId: updatePermission
      parameters:
        - name: id
          required: true
          in: path
          description: Permission ID (UUID)
          schema:
            example: 123e4567-e89b-12d3-a456-426614174000
            type: string
        - name: Authorization
          in: header
          description: Bearer auth token
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePermissionRequestDto"
      responses:
        "200":
          description: Update an existing permission by its ID.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PermissionWithRolesResponseDto"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                example:
                  statusCode: 400
                  message:
                    - field: field
                      message: message
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                example:
                  statusCode: 401
                  message: Unauthorized.
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                example:
                  statusCode: 403
                  message: Forbidden.
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                example:
                  statusCode: 404
                  message: Not Found.
        "422":
          description: Unprocessable Entity
          content:
            application/json:
              schema:
                example:
                  statusCode: 422
                  message: Unprocessable content.
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                example:
                  statusCode: 500
                  message: Internal server error.
      summary: Update a permission
      tags: *a3
    delete:
      operationId: deletePermission
      parameters:
        - name: id
          required: true
          in: path
          description: Permission ID (UUID)
          schema:
            example: 123e4567-e89b-12d3-a456-426614174000
            type: string
        - name: Authorization
          in: header
          description: Bearer auth token
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeletePermissionRequestDto"
      responses:
        "200":
          description: Delete a specific permission by its ID.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PermissionWithRolesResponseDto"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                example:
                  statusCode: 400
                  message:
                    - field: field
                      message: message
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                example:
                  statusCode: 401
                  message: Unauthorized.
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                example:
                  statusCode: 403
                  message: Forbidden.
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                example:
                  statusCode: 404
                  message: Not Found.
        "422":
          description: Unprocessable Entity
          content:
            application/json:
              schema:
                example:
                  statusCode: 422
                  message: Unprocessable content.
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                example:
                  statusCode: 500
                  message: Internal server error.
      summary: Delete a permission
      tags: *a3
...
info:
  title: Ecom API
  description: The Ecom API description
  version: "1.0"
  contact: {}
...
servers: []
components:
  schemas:
...
    PaginationResponseDto:
      type: object
      properties:
        totalPages:
          type: number
          description: Total number of pages
          example: 5
        totalItems:
          type: number
          description: Total number of items
          example: 100
        pageSize:
          type: number
          description: Number of items per page
          example: 20
        pageIndex:
          type: number
          description: Current page index (starts from 0)
          example: 0
      required:
        - totalPages
        - totalItems
        - pageSize
        - pageIndex
    PageDto:
      type: object
      properties:
        pagination:
          $ref: "#/components/schemas/PaginationResponseDto"
      required:
        - pagination
...
    RoleResponseDto:
      type: object
      properties:
        id:
          type: string
          description: Role ID
          example: 123e4567-e89b-12d3-a456-426614174000
        name:
          type: string
          description: Role name
          example: Admin
        isActive:
          type: boolean
          description: Is role active
          example: true
        description:
          type: string
          description: Role description
          example: Administrator role
      required:
        - id
        - name
        - isActive
    PermissionWithRolesResponseDto:
      type: object
      properties:
        id:
          type: string
          description: Permission ID
          example: 123e4567-e89b-12d3-a456-426614174000
        name:
          type: string
          description: Permission name
          example: Create User
        description:
          type: string
          description: Permission description
          example: Allows creating new users
        path:
          type: string
          description: Permission path
          example: /users
        method:
          type: string
          description: Permission method
          example: POST
        module:
          type: string
          description: Permission module
          example: USER
        roles:
          description: Roles associated with the permission
          example:
            - id: 123e4567-e89b-12d3-a456-426614174000
              name: Admin
              description: Administrator role
          type: array
          items:
            $ref: "#/components/schemas/RoleResponseDto"
      required:
        - id
        - name
        - path
        - method
        - module
    CreatePermissionRequestDto:
      type: object
      properties:
        name:
          type: string
          description: Permission name
          example: Create User
          minLength: 2
          maxLength: 500
        description:
          type: string
          description: Permission description
          example: Allows creating new users
          maxLength: 1000
        path:
          type: string
          description: Permission path
          example: /users
          minLength: 1
          maxLength: 1000
        method:
          type: string
          description: Permission method
          example: POST
          enum: &a15
            - GET
            - POST
            - PUT
            - DELETE
            - PATCH
        rolesIds:
          description: Roles associated with the permission
          example: &a16
            - 123e4567-e89b-12d3-a456-426614174000
            - 223e4567-e89b-12d3-a456-426614174001
          type: array
          items:
            type: string
      required:
        - name
        - path
        - method
        - rolesIds
    UpdatePermissionRequestDto:
      type: object
      properties:
        name:
          type: string
          description: Permission name
          example: Create User
          minLength: 2
          maxLength: 500
        description:
          type: string
          description: Permission description
          example: Allows creating new users
          maxLength: 1000
        path:
          type: string
          description: Permission path
          example: /users
          minLength: 1
          maxLength: 1000
        method:
          type: string
          description: Permission method
          example: POST
          enum: *a15
        rolesIds:
          description: Roles associated with the permission
          example: *a16
          type: array
          items:
            type: string
      required:
        - name
        - path
        - method
        - rolesIds
    DeletePermissionRequestDto:
      type: object
      properties:
        isHardDelete:
          type: boolean
          description: Whether to hard delete the permission
          example: false
          default: false
...